Issue #51: Add Http GET request method for server. Add exception handling. Restructure 'springboot' module
Added Http GET request method on REST API Spring Boot server for a list of Vending Machines.
Added ResourceNotFoundException and a handling for this by creating an ErrorDetails object. The object will be returned on Http request along with Http status code.
Restructured the 'springboot' module to follow MVC pattern. Necessary changes and refactoring of 'jsonio' and 'core' module as a result of this.