Issue #72: Create MachineTrackerAccesLocal class for accessing file directly
If the REST API server is not available, the application should be able to access the file directly using the same methods as with remote access.
The methods of the Access interace has been implemented with both business logic and exception throwing, as well as methods for accessing the file and doing necessary operations.
The intended error handling is throwing exceptions, which can be caught by the method's caller, i.e. controller classes, which can display the human-readable error messages.
Javadoc comments are also added.