Issue #73: Add HealthController class to Spring Boot. Edit UI controller to use both remote and local access
The UI controller has been edited to check the health of the REST API server (i.e. sending a GET request and check if it's available)
before choosing to use remote access. If the server is unavailable, the controller will use local access.

The health check runs in a seperate thread to ensure the UI remains responsive and the application doesn't freeze.
The list of vending machines will be updated after the task of checking the health is done.