Issue #75: Create AccessService class. Update  UI controller class
In order for the different UI controllers to share the same Access object, an AccessService class has been created to handle this as a shared resource.
To reflect the changes in how the Access object is acccessed, the main UI controller class has been updated to reflect these new changes.