Issue #77: Update findID type to String
In order to include the machineID "0" to be a valid ID, I changed the type of the method findID to be a String. Then the method can return null instead of "0" if no machine is selected. Then I had to parse findID to int to use it in other methods.