Issue #98: Enhamcements in refill scene and user input validation
Implemented a method to set the inventory, and added this
functionality to the method triggered during scene transition to ensure the inventory is displayed. In addition, the refillItem method has been enhanced to validate if the item insertion textField is empty. If it is, an error message will be displayed.