Issue #104: Add Maven Shade and Jpackage to build executable for Springboot server
The combination of Maven Shade plugin and Jpackage plugin is used to create platform-specific executables such that the client does not need to have Java installed.
Maven Shade plugin builds a shaded jar file with all the dependencies in a single jar file.
Jpackage plugin builds a platform-specific executable file, i.e. .exe for Windows, .dmg for MacOS and .deb for Linux.